[SP-313] #comment sonarqube 대응 23.05.19 #close #time 1h +review SR @dongmin.lee
[ Backend-Core :: PLE-TreeFramework :: 313DEVGRP ]
[Document] http://www.313.co.kr/confluence
[IssueTracker] http://www.313.co.kr/jira
[VersionControl] https://github.com/313devgrp
[CodeReview] http://www.313.co.kr/fecru
[CI-Build] http://www.313.co.kr/bamboo
[BuildManager] http://www.313.co.kr/spinnaker
[CD-Deploy] http://www.313.co.kr/jenkins
[ArtifactManager] http://www.313.co.kr/nexus
[CodeAnalysis] http://www.313.co.kr/sonar
[Docker] https://hub.docker.com/u/313devgrp

+++ src/main/java/com/egovframework/javaservice/treeframework/service/TreeServiceImpl.java
@@ -85,7 +85,7 @@ public class TreeServiceImpl implements TreeService {
         treeDao.getCurrentSession().setCacheMode(CacheMode.IGNORE);
         int totalCount = treeDao.getCount(treeSearchEntity);
 
-        double calPageSize = Math.ceil(totalCount / treeSearchEntity.getPageUnit());
+        double calPageSize = Math.ceil((double) totalCount / treeSearchEntity.getPageUnit());
         int autoPageSize = (int) Math.round(calPageSize);
 
         /** paging */
